extra confirmation on delete non-created chats when space owner

+++ src/components/Sidebar/Chat.js
@@ -56,6 +56,13 @@ const Chat = (props) => {
 
   const handleDelete = async (e) => {
     e.stopPropagation();
+    
+    if (
+      chat.type === 'public' && !isCreator && isSpaceOwner &&
+      !window.confirm(`Are you sure you want to delete ${chat.title}, created by another member?`)
+    ) {
+      return;
+    }
 
     try {
       await service.remove(`/chats/${chat.id}`);
